Updated the build system

+++ tool/build.ps1
@@ -2,9 +2,9 @@
 Set-StrictMode -Version Latest
 Set-Location (Split-Path $PSScriptRoot)
 
-Remove-Item src -Recurse
+if (Test-Path src) { Remove-Item src -Recurse }
 haxe build.hxml
 
-foreach ($item in "index.php", "src/Main.php") {
+foreach ($item in "index.php", "src/Main.php", "src/tink/cli/*0.php") {
 	if (Test-Path $item) { Remove-Item $item }
 }
